# src/controllers/contacts.js
import * as contactServices from '../services/contacts.js';
import createHttpError from 'http-errors';
import { parsePaginationParams } from '../utils/parsePaginationParams.js';
import { parseSortParams } from '../utils/parseSortParams.js';
import { sortByList } from '../db/models/Contact.js';
import { parseContactFilterParams } from '../utils/parseContactFilterParams.js';
import { saveFileToUploadDir } from '../middlewares/saveFileToUploadDir.js';
import * as path from "node:path";
import {env} from "../utils/env.js";
import { saveFileToCloudinary } from '../utils/saveFileToCloudinary.js';

const enableCloudinary = env("ENABLE_CLOUDINARY");

export const getContactsController = async (req, res) => {
  const { page, perPage } = parsePaginationParams(req.query);
  const { sortBy, sortOrder } = parseSortParams(req.query, sortByList);
  const isFavouriteFilter = parseContactFilterParams(req.query);
  const typeFilter = req.query.contactType;
  const {_id: userIdFilter} = req.user;
  const data = await contactServices.getContacts({
    page,
    perPage,
    sortBy,
    sortOrder,
    isFavouriteFilter,
    typeFilter,
    userIdFilter
  });

  // if (!data) {
  //   throw createHttpError(404, `Contact not found`);
  // }
  res.json({
    status: 200,
    message: 'Successfully found contacts!',
    data,
  });
};

export const getContactByIdController = async (req, res, next) => {
  const { contactId } = req.params;
  console.log(contactId);
  const data = await contactServices.getContactById(contactId);

  if (!data) {
    throw createHttpError(404, `Contact not found`);
  }
  res.json({
    status: 200,
    message: `Successfully found contact!`,
    data,
  });
};

export const addContactController = async (req, res) => {

const {_id: userId} = req.user;
let photo = null;
if(req.file){
  if(enableCloudinary === "true"){
    photo = await saveFileToCloudinary(req.file, "photo");
  }else{
     await saveFileToUploadDir(req.file);
  photo = path.join(req.file.filename);
  }
 
}
  const data = await contactServices.addContact({...req.body, photo, userId});

  res.status(201).json({
    status: 201,
    message: 'Successfully created a contact!',
    data,
  });
};

export const upsertContactController = async (req, res) => {
  const { contactId: _id } = req.params;
  //const _id = req.params.id;

  const result = await contactServices.updateContact({
    _id,
    payload: req.body,
    options: { upsert: true },
  });

  const status = result.isNew ? 201 : 200;
  res.status(status).json({
    status,
    message: result.isNew
      ? 'Successfully created a contact!'
      : 'Successfully patched a contact!',
    data: result.data,
  });
};

export const patchContactController = async (req, res) => {
  const { contactId: _id } = req.params;
  const result = await contactServices.updateContact({
    _id,
    payload: req.body,
  });
  if (!result) {
    throw createHttpError(404, `Contact not found`);
  }
  res.json({
    status: 200,
    message: `Successfully patched a contact!`,
    data: result.data,
  });
};

export const deleteContactController = async (req, res) => {
  const { contactId: _id } = req.params;
  const data = await contactServices.deleteContact({ _id });
  if (!data) {
    throw createHttpError(404, `Contact not found`);
  }
  res.status(204).send();
};